requirements.txt
ipython==8.12.0
matplotlib==3.7.1
numpy==1.24.2
pydantic==1.10.7
pystoi==0.3.3
pyyaml==6.0
scipy==1.10.1
soundfile==0.12.1
torch==2.0.0
torchaudio==2.0.1
torchvision==0.15.1

# Install pypesq after install all others stuffs (numpy, ...)
librosa==0.10.0.post2
pypesq==1.2.4